pyproject.toml
[project]
name = "minimal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.30",
    "duckduckgo-search>=7.5.1",
    "litellm>=1.61.11",
    "markdownify>=1.1.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "smolagents>=1.10.0",
    "tavily-python>=0.7.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/minimal_agent"]
